+++ packages/@n8n/local-gateway/src/main/notifications.ts
@@ -17,6 +17,8 @@ export interface PromptNotifierDeps {
 export interface PromptNotifier {
 	notifyLocalPrompt(prompt: LocalPermissionPromptRequest): void;
 	notifyConfirmationRequest(payload: InstanceAiConfirmationRequestPayload): void;
+	/** A one-shot task run resolved (done/handoff/error). Strings come localized from the renderer. */
+	notifyTaskResult(title: string, body: string): void;
 }
 
 /**
@@ -46,6 +48,10 @@ export function createPromptNotifier(deps: PromptNotifierDeps): PromptNotifier {
 			if (notifiedRequestIds.size >= MAX_REMEMBERED_REQUEST_IDS) notifiedRequestIds.clear();
 			notifiedRequestIds.add(payload.requestId);
 			show('AI Assistant needs your input', payload.message);
+		},
+
+		notifyTaskResult(title, body) {
+			show(title, body);
 		},
 	};
 }
