+++ packages/@n8n/local-gateway/src/main/preload.ts
@@ -201,6 +201,10 @@ const electronApi: ElectronApi = {
 		await ipcRenderer.invoke('permissions:openSettings', kind);
 	},
 
+	notifyTaskResult: async (title: string, body: string): Promise<void> => {
+		await ipcRenderer.invoke('notifications:taskResult', title, body);
+	},
+
 	listPermissionPrompts: async (): Promise<LocalPermissionPromptRequest[]> =>
 		await (ipcRenderer.invoke('permissionPrompt:list') as Promise<LocalPermissionPromptRequest[]>),
