+++ packages/@n8n/local-gateway/src/renderer/App.vue
@@ -4,6 +4,7 @@ import { computed, onMounted, ref, watch } from 'vue';
 import AppHeader from './components/AppHeader.vue';
 import ChatPanel from './components/ChatPanel.vue';
 import PermissionPromptStack from './components/PermissionPromptStack.vue';
+import TaskResultCard from './components/TaskResultCard.vue';
 import ComplexTaskView from './views/ComplexTaskView.vue';
 import HomeView from './views/HomeView.vue';
 import SettingsView from './views/SettingsView.vue';
@@ -12,7 +13,9 @@ import TaskDetailView from './views/TaskDetailView.vue';
 import TaskDraftView from './views/TaskDraftView.vue';
 import TaskSetupView from './views/TaskSetupView.vue';
 
+import { dismissTaskResult, taskResultState } from './assistant/task-result-store';
 import { useAssistantScreen } from './assistant/use-assistant-screen';
+import { usePendingTasks } from './assistant/use-pending-tasks';
 import { chatOverlay, closeChat } from './chat/chat-overlay';
 import { clearAllPrompts } from './permissions/permission-prompt-store';
 import { getThreadPromptWatcher } from './permissions/thread-prompt-watcher';
@@ -28,6 +31,17 @@ const auth = ref<AuthStatus>({
 });
 const { screen, goHome } = useAssistantScreen();
 const { prompts, respondingIds, failedIds, respondToPrompt } = usePermissionPrompts();
+const pendingTasks = usePendingTasks();
+
+/** Keep the one-off: promote its thread into a saved task; home shows the pending entry. */
+function keepTaskResult() {
+	const card = taskResultState.card;
+	if (card?.kind !== 'done') return;
+	// The icon (from the agent's outcome report) lands on the workflow's meta, not in its name.
+	pendingTasks.promote(card.threadId, card.label, card.icon);
+	dismissTaskResult();
+	closeChat();
+}
 
 // The setup and complex screens carry their own back-header, so the main
 // AppHeader is suppressed for them; home and draft keep it.
@@ -62,6 +76,7 @@ watch(
 			showSettings.value = false;
 			goHome();
 			closeChat();
+			dismissTaskResult();
 			// Another user must never see this user's pending prompts.
 			getThreadPromptWatcher().stopAllWatches();
 			clearAllPrompts();
@@ -110,16 +125,24 @@ watch(
 				</div>
 			</Transition>
 
-			<!-- Permission prompts float above the active composer, on any view. -->
+			<!-- Permission prompts and the task result card float above the active
+			     composer, on any view — including over the chat overlay. -->
 			<div
-				v-if="auth.state === 'signedIn' && prompts.length > 0"
+				v-if="auth.state === 'signedIn' && (prompts.length > 0 || taskResultState.card)"
 				:class="[$style.promptArea, chatOverlay.isOpen ? $style.aboveChat : $style.aboveHome]"
 			>
 				<PermissionPromptStack
+					v-if="prompts.length > 0"
 					:prompts="prompts"
 					:responding-ids="respondingIds"
 					:failed-ids="failedIds"
 					@respond="respondToPrompt"
+				/>
+				<TaskResultCard
+					v-if="taskResultState.card"
+					:card="taskResultState.card"
+					@keep="keepTaskResult"
+					@dismiss="dismissTaskResult"
 				/>
 			</div>
 		</div>
@@ -169,6 +192,9 @@ watch(
 	right: var(--spacing--xs);
 	left: var(--spacing--xs);
 	z-index: 10;
+	display: flex;
+	flex-direction: column;
+	gap: var(--spacing--2xs);
 }
 
 /* Clears the chat panel's single-row composer. */
