+++ packages/@n8n/local-gateway/src/renderer/components/TaskComposer.vue
@@ -1,50 +1,37 @@
 <script setup lang="ts">
 import { N8nIcon, N8nTooltip } from '@n8n/design-system';
 import { useI18n } from '@n8n/i18n';
-import { computed, nextTick, onMounted, ref, watch } from 'vue';
+import { computed, onMounted, ref, watch } from 'vue';
 
-import AssistantButton from './AssistantButton.vue';
 import ComposerField from './ComposerField.vue';
 import ContextPill from './ContextPill.vue';
 import MiniSpinner from './MiniSpinner.vue';
 
 import { suggestionChipsFor } from '../assistant/contexts';
 import { watchAssistantRun } from '../assistant/run-watcher';
+import {
+	dismissTaskResult,
+	showTaskResult,
+	type TaskResultCard,
+} from '../assistant/task-result-store';
 import { openChat } from '../chat/chat-overlay';
 import { useAssistantContext } from '../assistant/use-assistant-context';
-import { usePendingTasks } from '../assistant/use-pending-tasks';
 import { getThreadPromptWatcher } from '../permissions/thread-prompt-watcher';
 import type { DesktopAssistantTaskRequest } from '../../shared/types';
 
 type ComposerState = 'idle' | 'thinking' | 'doing';
 
-/**
- * The card shown above the input once a run finishes:
- * - `done`: the assistant did the thing — offer to keep it as a saved task.
- *   `label` is the agent's outcome title (or the truncated prompt) and doubles
- *   as the workflow name when the task is kept.
- * - `handoff`: the agent reported a failed task outcome — point the user to
- *   the instance UI; `message` is the agent's failure reason when it gave one.
- * - `error`: the run errored, timed out, or was canceled.
- *
- * A run that finishes with no outcome and no action shows no card at all: the
- * agent replied in text, so the thread is opened in the chat overlay instead.
- */
-type ResultCard =
-	| { kind: 'done'; threadId: string; label: string; icon?: string }
-	| { kind: 'handoff'; message?: string }
-	| { kind: 'error'; message: string };
+// The result card itself floats at App level (task-result-store) so it stays
+// visible over the chat overlay. A run that finishes with no outcome and no
+// action shows no card at all: the agent replied in text, so the thread is
+// opened in the chat overlay instead.
 
 const MAX_LABEL_LENGTH = 60;
 
 const i18n = useI18n();
-const pendingTasks = usePendingTasks();
-
-const emit = defineEmits<{ kept: [] }>();
 
 const text = ref('');
 const state = ref<ComposerState>('idle');
-const resultCard = ref<ResultCard | null>(null);
 
 // Context selection is shared app-wide (see use-assistant-context). `detected` is
 // the chosen raw context we forward to the backend; `activeContext`/`contextOptions`
@@ -73,9 +60,6 @@ const screenshotTooltip = computed(() =>
 
 onMounted(ensureDetection);
 
-const inputRef = ref<InstanceType<typeof ComposerField> | null>(null);
-const resultCardRef = ref<HTMLElement | null>(null);
-
 const busy = computed(() => state.value === 'thinking' || state.value === 'doing');
 
 const chips = computed(() =>
@@ -86,11 +70,28 @@ function truncateLabel(value: string): string {
 	return value.length > MAX_LABEL_LENGTH ? `${value.slice(0, MAX_LABEL_LENGTH - 1)}…` : value;
 }
 
-function showResult(card: ResultCard) {
-	resultCard.value = card;
-	// The card is interactive and appears far from the input, so move focus to
-	// it; the live region still announces it for screen readers.
-	void nextTick(() => resultCardRef.value?.focus());
+/**
+ * Store the result card (it floats at App level) and mirror it as a system
+ * notification — the main process shows it only while the window is hidden,
+ * and clicking it brings the window (with the card) back up.
+ */
+function showResult(card: TaskResultCard) {
+	showTaskResult(card);
+	const [title, body] =
+		card.kind === 'done'
+			? [
+					`${i18n.baseText('desktopAssistant.composer.done')}: ${card.label}`,
+					i18n.baseText('desktopAssistant.composer.keepPrompt'),
+				]
+			: card.kind === 'handoff'
+				? [
+						i18n.baseText('desktopAssistant.composer.handoffTitle'),
+						card.message ?? i18n.baseText('desktopAssistant.composer.handoffMessage'),
+					]
+				: [i18n.baseText('desktopAssistant.composer.errorTitle'), card.message];
+	window.electronAPI.notifyTaskResult(title, body).catch((error: unknown) => {
+		console.error('Failed to show the task result notification', error);
+	});
 }
 
 /**
@@ -130,7 +131,7 @@ async function submit(prompt?: string) {
 	if (!value || busy.value) return;
 
 	text.value = '';
-	resultCard.value = null;
+	dismissTaskResult();
 	state.value = 'thinking';
 
 	try {
@@ -190,28 +191,6 @@ async function submit(prompt?: string) {
 	}
 }
 
-// Return focus to the input once the result card goes away, so keyboard users
-// aren't stranded on the removed card.
-function returnFocusToInput() {
-	void nextTick(() => inputRef.value?.focus());
-}
-
-/** Keep the one-off: promote its thread into a saved task and show the Tasks list. */
-function keepTask() {
-	if (resultCard.value?.kind !== 'done') return;
-	const { threadId, label, icon } = resultCard.value;
-	// The icon (from the agent's outcome report) lands on the workflow's meta, not in its name.
-	pendingTasks.promote(threadId, label, icon);
-	resultCard.value = null;
-	returnFocusToInput();
-	emit('kept');
-}
-
-function dismissResult() {
-	resultCard.value = null;
-	returnFocusToInput();
-}
-
 // Exposed so sibling surfaces (e.g. an empty-state recommendation card) can fire
 // a prompt through the same path as typing or clicking a suggestion chip.
 defineExpose({ submit });
@@ -222,8 +201,8 @@ defineExpose({ submit });
 		<div :class="$style.fieldArea">
 			<!--
 				Persistent live region: it stays mounted across states so screen
-				readers announce the Thinking/Doing pill and the result cards when
-				their content changes, rather than missing the insertion of a fresh node.
+				readers announce the Thinking/Doing pill when its content changes,
+				rather than missing the insertion of a fresh node.
 			-->
 			<div
 				role="status"
@@ -239,111 +218,10 @@ defineExpose({ submit });
 							: i18n.baseText('desktopAssistant.composer.thinking')
 					}}</span>
 				</div>
-
-				<!-- Done card -->
-				<div
-					v-if="resultCard?.kind === 'done'"
-					ref="resultCardRef"
-					tabindex="-1"
-					:class="[$style.resultCard, $style.doneCard]"
-					data-testid="composer-done-card"
-				>
-					<div :class="[$style.resultHeader, $style.doneHeader]">
-						<span :class="[$style.resultBadge, $style.doneBadge]">
-							<N8nIcon icon="check" :size="13" aria-hidden="true" />
-						</span>
-						<span :class="[$style.resultTitle, $style.doneTitle]">{{
-							i18n.baseText('desktopAssistant.composer.done')
-						}}</span>
-						<span :class="$style.resultSubtitle">{{ resultCard.label }}</span>
-						<button
-							type="button"
-							:class="$style.resultDismiss"
-							:aria-label="i18n.baseText('desktopAssistant.composer.dismiss')"
-							@click="dismissResult"
-						>
-							<N8nIcon icon="x" :size="14" aria-hidden="true" />
-						</button>
-					</div>
-					<div :class="$style.resultBody">
-						<div :class="$style.resultMessage">
-							{{ i18n.baseText('desktopAssistant.composer.keepPrompt') }}
-						</div>
-						<div :class="$style.resultActions">
-							<AssistantButton @click="dismissResult">
-								{{ i18n.baseText('desktopAssistant.composer.noThanks') }}
-							</AssistantButton>
-							<AssistantButton variant="solid" @click="keepTask">
-								{{ i18n.baseText('desktopAssistant.composer.saveAsReady') }}
-							</AssistantButton>
-						</div>
-					</div>
-				</div>
-
-				<!-- Handoff card -->
-				<div
-					v-if="resultCard?.kind === 'handoff'"
-					ref="resultCardRef"
-					tabindex="-1"
-					:class="[$style.resultCard, $style.handoffCard]"
-					data-testid="composer-handoff-card"
-				>
-					<div :class="[$style.resultHeader, $style.handoffHeader]">
-						<span :class="[$style.resultBadge, $style.handoffBadge]">
-							<N8nIcon icon="info" :size="13" aria-hidden="true" />
-						</span>
-						<span :class="[$style.resultTitle, $style.handoffTitle]">{{
-							i18n.baseText('desktopAssistant.composer.handoffTitle')
-						}}</span>
-						<button
-							type="button"
-							:class="$style.resultDismiss"
-							:aria-label="i18n.baseText('desktopAssistant.composer.dismiss')"
-							@click="dismissResult"
-						>
-							<N8nIcon icon="x" :size="14" aria-hidden="true" />
-						</button>
-					</div>
-					<div :class="$style.resultBody">
-						<div :class="$style.resultMessage">
-							{{ resultCard.message ?? i18n.baseText('desktopAssistant.composer.handoffMessage') }}
-						</div>
-					</div>
-				</div>
-
-				<!-- Error card -->
-				<div
-					v-if="resultCard?.kind === 'error'"
-					ref="resultCardRef"
-					tabindex="-1"
-					:class="[$style.resultCard, $style.errorCard]"
-					data-testid="composer-error-card"
-				>
-					<div :class="[$style.resultHeader, $style.errorHeader]">
-						<span :class="[$style.resultBadge, $style.errorBadge]">
-							<N8nIcon icon="triangle-alert" :size="13" aria-hidden="true" />
-						</span>
-						<span :class="[$style.resultTitle, $style.errorTitle]">{{
-							i18n.baseText('desktopAssistant.composer.errorTitle')
-						}}</span>
-						<button
-							type="button"
-							:class="$style.resultDismiss"
-							:aria-label="i18n.baseText('desktopAssistant.composer.dismiss')"
-							@click="dismissResult"
-						>
-							<N8nIcon icon="x" :size="14" aria-hidden="true" />
-						</button>
-					</div>
-					<div :class="$style.resultBody">
-						<div :class="$style.resultMessage">{{ resultCard.message }}</div>
-					</div>
-				</div>
 			</div>
 
 			<!-- Input field -->
 			<ComposerField
-				ref="inputRef"
 				v-model="text"
 				:disabled="busy"
 				:busy="busy"
@@ -416,156 +294,6 @@ defineExpose({ submit });
 	border: 1px solid var(--da-border);
 	border-radius: var(--radius--full);
 	box-shadow: 0 6px 18px rgba(0, 0, 0, 0.35);
-}
-
-/* Shared result-card shell; the done/handoff/error variants only swap accents. */
-.resultCard {
-	position: absolute;
-	right: 0;
-	bottom: calc(100% + var(--spacing--2xs));
-	left: 0;
-	overflow: hidden;
-	background: var(--da-surface-2);
-	border-radius: var(--da-radius);
-	box-shadow: 0 8px 24px rgba(0, 0, 0, 0.45);
-}
-
-/* The card receives programmatic focus when it appears; suppress the ring there
-   (it's an announcement target, not an interactive control). */
-.resultCard:focus {
-	outline: none;
-}
-
-.resultHeader {
-	display: flex;
-	align-items: center;
-	gap: var(--spacing--2xs);
-	padding: 10px var(--spacing--xs);
-}
-
-.resultBadge {
-	display: inline-flex;
-	flex-shrink: 0;
-	align-items: center;
-	justify-content: center;
-	width: 20px;
-	height: 20px;
-	border-radius: 50%;
-}
-
-.resultTitle {
-	font-size: 13px;
-	font-weight: 600;
-}
-
-.resultSubtitle {
-	overflow: hidden;
-	font-size: 12px;
-	color: var(--da-subtler);
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.resultDismiss {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	width: 24px;
-	height: 24px;
-	margin-left: auto;
-	color: var(--da-subtler);
-	cursor: pointer;
-	background: none;
-	border: none;
-	border-radius: 7px;
-	transition:
-		background 0.12s,
-		color 0.12s;
-}
-
-.resultDismiss:hover {
-	color: var(--da-text);
-	background: var(--da-surface);
-}
-
-.resultDismiss:focus-visible {
-	color: var(--da-text);
-	background: var(--da-surface);
-	outline: var(--da-focus-ring);
-	outline-offset: var(--da-focus-ring-offset);
-}
-
-.resultBody {
-	padding: var(--spacing--xs);
-}
-
-.resultMessage {
-	font-size: 12px;
-	color: var(--da-subtler);
-}
-
-.resultActions {
-	display: flex;
-	gap: var(--spacing--2xs);
-	justify-content: space-between;
-	margin-top: var(--spacing--xs);
-}
-
-/* Done variant: green accents. */
-.doneCard {
-	border: 1px solid rgba(63, 207, 142, 0.45);
-}
-
-.doneHeader {
-	background: rgba(63, 207, 142, 0.12);
-	border-bottom: 1px solid rgba(63, 207, 142, 0.22);
-}
-
-.doneBadge {
-	color: var(--da-green);
-	background: rgba(63, 207, 142, 0.2);
-}
-
-.doneTitle {
-	color: var(--da-green);
-}
-
-/* Handoff variant: amber accents — informational, not a failure. */
-.handoffCard {
-	border: 1px solid rgba(245, 184, 75, 0.45);
-}
-
-.handoffHeader {
-	background: rgba(245, 184, 75, 0.12);
-	border-bottom: 1px solid rgba(245, 184, 75, 0.22);
-}
-
-.handoffBadge {
-	color: var(--da-amber);
-	background: rgba(245, 184, 75, 0.2);
-}
-
-.handoffTitle {
-	color: var(--da-amber);
-}
-
-/* Error variant: red accents. */
-.errorCard {
-	border: 1px solid rgba(255, 107, 107, 0.45);
-}
-
-.errorHeader {
-	background: rgba(255, 107, 107, 0.12);
-	border-bottom: 1px solid rgba(255, 107, 107, 0.22);
-}
-
-.errorBadge {
-	color: var(--da-red);
-	background: rgba(255, 107, 107, 0.2);
-}
-
-.errorTitle {
-	color: var(--da-red);
 }
 
 .pillRow {
